Update Provisioning-Deployment.org
Running `npm install` assumes that there is the file `package.json` in the working directory. Since this is not the case, the image build fails. This fixes the error. Alternatively, run `npm init -y` before `npm install`, but this is cleaner.

Or was the intention to let the students find out the solution? Quite frustrating though if you never have experienced docker and its error messages...